Refactor as Prop3D.rotation_from and add SciPy support
