+++ pyvista/core/utilities/arrays.py
@@ -13,7 +13,6 @@ from typing import Union
 import numpy as np
 
 import pyvista
-from pyvista.core import _validation
 from pyvista.core import _vtk_core as _vtk
 from pyvista.core.errors import AmbiguousDataError
 from pyvista.core.errors import MissingDataError
@@ -935,104 +934,3 @@ class _SerializedDictArray(UserDict, _vtk.vtkStringArray):  # type: ignore[type-
     def setdefault(self, *args, **kwargs):
         super().setdefault(*args, **kwargs)
         self._update_string()
-
-
-def rotation_matrix_to_orientation(
-    array: NumpyArray[float] | _vtk.vtkMatrix3x3,
-) -> tuple[float, float, float]:
-    """Convert a 3x3 rotation matrix to x-y-z orientation angles.
-
-    The orientation angles define rotations about the world's x-y-z axes. The angles
-    are specified in degrees and in x-y-z order. However, the rotations should
-    be applied in the order: first rotate about the y-axis, then x-axis, then z-axis.
-
-    The rotation angles and rotation matrix can be used interchangeably for
-    transformations.
-
-    Parameters
-    ----------
-    array : NumpyArray[float] | vtkMatrix3x3
-        3x3 rotation matrix as a NumPy array or a vtkMatrix.
-
-    Returns
-    -------
-    tuple
-        Tuple with x-y-z axis rotation angles in degrees.
-
-    See Also
-    --------
-    rotation_matrix_to_orientation_angles
-        Inverse function for this conversion.
-
-    pyvista.DataSet.rotate_x, pyvista.DataSet.rotate_y, pyvista.DataSet.rotate_z, pyvista.DataSet.transform
-        Relevant methods for transforming datasets.
-
-    pyvista.Prop3D.orientation, pyvista.Prop3D.rotation
-        Relevant methods for transforming 3D props or :class:`pyvista.Actor`.
-
-    Examples
-    --------
-    >>> import pyvista as pv
-    >>> pv.rotation_matrix_to_orientation(
-    ...     np.array([[0.5, 0.5, 0], [0.5, 0.5, 0], [0, 0, 1]])
-    ... )
-    (0.0, -0.0, 90.0)
-    """
-    array_3x3 = _validation.validate_transform3x3(array)
-    array_4x4 = np.eye(4)
-    array_4x4[:3, :3] = array_3x3
-    transform = _vtk.vtkTransform()
-    transform.SetMatrix(array_4x4.ravel())
-    return transform.GetOrientation()
-
-
-def orientation_to_rotation_matrix(orientation: VectorLike[float]) -> NumpyArray[float]:
-    """Convert x-y-z orientation angles to a 3x3 matrix.
-
-    The orientation angles define rotations about the world's x-y-z axes. The angles
-    are specified in degrees and in x-y-z order. However, the rotations should
-    be applied in the order: first rotate about the y-axis, then x-axis, then z-axis.
-
-    The rotation angles and rotation matrix can be used interchangeably for
-    transformations.
-
-    Parameters
-    ----------
-    orientation : VectorLike[float]
-        The x-y-z axis orientation angles in degrees.
-
-    Returns
-    -------
-    numpy.ndarray
-        3x3 rotation matrix.
-
-    See Also
-    --------
-    rotation_matrix_to_orientation_angles
-        Inverse function for this conversion.
-
-    pyvista.DataSet.rotate_x, pyvista.DataSet.rotate_y, pyvista.DataSet.rotate_z, pyvista.DataSet.transform
-        Relevant methods for transforming datasets.
-
-    pyvista.Prop3D.orientation, pyvista.Prop3D.rotation
-        Relevant methods for transforming 3D props or :class:`pyvista.Actor`.
-
-    Examples
-    --------
-    >>> import pyvista as pv
-    >>> pv.orientation_to_rotation_matrix((10, 20, 30))
-    array([[ 0.78410209, -0.49240388,  0.37778609],
-           [ 0.52128058,  0.85286853,  0.02969559],
-           [-0.33682409,  0.17364818,  0.92541658]])
-
-    """
-    # Local import - vtkActor is usually only available in plotting module
-    from vtk import vtkActor
-
-    valid_orientation = _validation.validate_array3(orientation)
-    # Use the transformations inherited from vtkProp3D class for the conversion
-    prop = vtkActor()
-    prop.SetOrientation(valid_orientation)
-    matrix = _vtk.vtkMatrix4x4()
-    prop.GetMatrix(matrix)
-    return array_from_vtkmatrix(matrix)[:3, :3]
