+++ pyvista/plotting/prop3d.py
@@ -6,12 +6,15 @@ from typing import TYPE_CHECKING
 
 import numpy as np
 
+from pyvista.core import _validation
 from pyvista.core.utilities.arrays import _coerce_transformlike_arg
 from pyvista.core.utilities.arrays import array_from_vtkmatrix
 from pyvista.core.utilities.arrays import vtkmatrix_from_array
 from pyvista.plotting import _vtk
 
 if TYPE_CHECKING:  # pragma: no cover
+    import scipy
+
     from pyvista.core._typing_core import BoundsLike
     from pyvista.core._typing_core import NumpyArray
     from pyvista.core._typing_core import TransformLike
@@ -193,6 +196,11 @@ class Prop3D(_vtk.vtkProp3D):
         and finally :func:`~rotate_z`.
 
         Rotations are applied about the specified :attr:`~origin`.
+
+        See Also
+        --------
+        rotation
+            Set the orientation This property may be used as an alternative to :attrs:`orientation`.
 
         Examples
         --------
@@ -376,3 +384,110 @@ class Prop3D(_vtk.vtkProp3D):
         1.7272069317100354
         """
         return self.GetLength()
+
+    def rotation_from(self, rotation: NumpyArray[float] | scipy.spatial.transform.Rotation):
+        """Set the entity's rotation.
+
+        Set the rotation of this entity from a 3x3 rotation matrix or a SciPy
+        ``Rotation`` object. This method may be used as an alternative for setting the
+        :attrs:`orientation`.
+
+        Parameters
+        ----------
+        rotation : NumpyArray[float] | scipy.spatial.transform.Rotation
+            3x3 rotation matrix or a SciPy ``Rotation`` object.
+
+        Examples
+        --------
+        Create an actor and show its initial orientation.
+
+        >>> import pyvista as pv
+        >>> pl = pv.Plotter()
+        >>> actor = pl.add_mesh(pv.Sphere())
+        >>> actor.orientation
+        (0.0, -0.0, 0.0)
+
+        Set the orientation using a 3x3 matrix.
+
+        >>> actor.rotation_from(
+        ...     np.array([[0, 1, 0], [1, 0, 0], [0, 0, 1]])
+        ... )
+        >>> actor.orientation
+        (0.0, -180.0, -89.99999999999999)
+
+
+        """
+        try:
+            orientation = _rotation_matrix_as_orientation(rotation)
+        except (TypeError, ValueError):
+            err_msg = f'Rotation must be a 3x3 NumPy array or a SciPy Rotation instance. Got {type(rotation)}.'
+            try:
+                from scipy.spatial.transform import Rotation
+            except ModuleNotFoundError:
+                raise TypeError(err_msg)
+            else:
+                if not isinstance(rotation, Rotation):
+                    raise TypeError(err_msg)
+                orientation = rotation.as_euler('yxz', degrees=True)
+
+        self.orientation = orientation
+
+
+def _rotation_matrix_as_orientation(
+    array: NumpyArray[float] | _vtk.vtkMatrix3x3,
+) -> tuple[float, float, float]:
+    """Convert a 3x3 rotation matrix to x-y-z orientation angles.
+
+    The orientation angles define rotations about the world's x-y-z axes. The angles
+    are specified in degrees and in x-y-z order. However, the rotations should
+    be applied in the order: first rotate about the y-axis, then x-axis, then z-axis.
+
+    The rotation angles and rotation matrix can be used interchangeably for
+    transformations.
+
+    Parameters
+    ----------
+    array : NumpyArray[float] | vtkMatrix3x3
+        3x3 rotation matrix as a NumPy array or a vtkMatrix.
+
+    Returns
+    -------
+    tuple
+        Tuple with x-y-z axis rotation angles in degrees.
+
+    """
+    array_3x3 = _validation.validate_transform3x3(array)
+    array_4x4 = np.eye(4)
+    array_4x4[:3, :3] = array_3x3
+    transform = _vtk.vtkTransform()
+    transform.SetMatrix(array_4x4.ravel())
+    return transform.GetOrientation()
+
+
+def _orientation_as_rotation_matrix(orientation: VectorLike[float]) -> NumpyArray[float]:
+    """Convert x-y-z orientation angles to a 3x3 matrix.
+
+    The orientation angles define rotations about the world's x-y-z axes. The angles
+    are specified in degrees and in x-y-z order. However, the rotations should
+    be applied in the order: first rotate about the y-axis, then x-axis, then z-axis.
+
+    The rotation angles and rotation matrix can be used interchangeably for
+    transformations.
+
+    Parameters
+    ----------
+    orientation : VectorLike[float]
+        The x-y-z axis orientation angles in degrees.
+
+    Returns
+    -------
+    numpy.ndarray
+        3x3 rotation matrix.
+
+    """
+    valid_orientation = _validation.validate_array3(orientation)
+    prop = _vtk.vtkActor()
+    prop.SetOrientation(valid_orientation)
+    matrix = _vtk.vtkMatrix4x4()
+    prop.GetMatrix(matrix)
+    return array_from_vtkmatrix(matrix)[:3, :3]
