+++ tests/plotting/test_actor.py
@@ -4,10 +4,13 @@ import platform
 
 import numpy as np
 import pytest
+import scipy
 import vtk
 
 import pyvista as pv
 from pyvista import examples
+from pyvista.plotting.prop3d import _orientation_as_rotation_matrix
+from pyvista.plotting.prop3d import _rotation_matrix_as_orientation
 
 skip_mac = pytest.mark.skipif(
     platform.system() == 'Darwin',
@@ -229,3 +232,41 @@ def test_vol_actor_prop(vol_actor):
     prop = vtk.vtkVolumeProperty()
     vol_actor.prop = prop
     assert vol_actor.prop is prop
+
+
+@pytest.mark.parametrize('func', [np.ndarray, scipy.spatial.transform.Rotation.from_matrix])
+def test_rotation_from(actor, func):
+    array = [
+        [0.78410209, -0.49240388, 0.37778609],
+        [0.52128058, 0.85286853, 0.02969559],
+        [-0.33682409, 0.17364818, 0.92541658],
+    ]
+
+    actor.rotation_from(array)
+
+    expected = (10, 20, 30)
+    actual = actor.orientation
+    assert np.allclose(actual, expected)
+
+
+@pytest.mark.parametrize('order', ['F', 'C'])
+def test_convert_orientation_to_rotation_matrix(order):
+    orientation = (10, 20, 30)
+    rotation = np.array(
+        [
+            [0.78410209, -0.49240388, 0.37778609],
+            [0.52128058, 0.85286853, 0.02969559],
+            [-0.33682409, 0.17364818, 0.92541658],
+        ],
+        order=order,
+    )
+
+    actual_rotation = _orientation_as_rotation_matrix(orientation)
+    assert isinstance(actual_rotation, np.ndarray)
+    assert actual_rotation.shape == (3, 3)
+    assert np.allclose(actual_rotation, rotation)
+
+    actual_orientation = _rotation_matrix_as_orientation(rotation)
+    assert isinstance(actual_orientation, tuple)
+    assert len(actual_orientation) == 3
+    assert np.allclose(actual_orientation, orientation)
